requirements.txt
boto3
botocore
requests
psycopg2-binary
python-dotenv